src/requirements.txt
anthropic==0.15.1
datasets==2.16.1
matplotlib==3.7.2
numpy==1.24.4
openai==1.12.0
pandas==2.2.0
protobuf==4.25.2
PyYAML==6.0.1
scikit_learn==1.3.0
torch==2.1.2
